requirements.txt
asgiref==3.8.1
Django==5.2.3
django-debug-toolbar==5.2.0
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
